# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sistema de Inventario Comercial API",
    "version": "1.0.0",
    "description": "API para la gestión de inventario, compras y ventas"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
    "/api/users": {
      "get": {
        "summary": "Obtener todos los usuarios",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Buscar por nombre o email"
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por rol"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sales": {
      "post": {
        "summary": "Registra una nueva venta con sus detalles",
        "tags": [
          "Sales"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "modoVenta",
                  "detalles"
                ],
                "properties": {
                  "modoVenta": {
                    "type": "string",
                    "enum": [
                      "local",
                      "viaje"
                    ]
                  },
                  "detalles": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "productoId",
                        "cantidad",
                        "precioCobrado"
                      ],
                      "properties": {
                        "productoId": {
                          "type": "string"
                        },
                        "cantidad": {
                          "type": "integer"
                        },
                        "precioCobrado": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Venta registrada exitosamente"
          },
          "400": {
            "description": "Stock insuficiente o datos inválidos"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/purchases": {
      "post": {
        "summary": "Registra un ingreso de mercadería (compra)",
        "tags": [
          "Purchases"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productoId",
                  "proveedorId",
                  "costoTotalCaja",
                  "unidadesPorCaja",
                  "cantidadCajas"
                ],
                "properties": {
                  "productoId": {
                    "type": "string"
                  },
                  "proveedorId": {
                    "type": "string"
                  },
                  "costoTotalCaja": {
                    "type": "number"
                  },
                  "unidadesPorCaja": {
                    "type": "integer"
                  },
                  "cantidadCajas": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Compra registrada exitosamente"
          },
          "400": {
            "description": "Error en los datos enviados"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/providers": {
      "get": {
        "summary": "Obtener todos los proveedores con paginación",
        "tags": [
          "Providers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Buscar por nombre"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de proveedores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Provider"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo proveedor",
        "tags": [
          "Providers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre"
                ],
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "telefono": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proveedor creado"
          }
        }
      }
    },
    "/api/providers/{id}": {
      "get": {
        "summary": "Obtener un proveedor por ID",
        "tags": [
          "Providers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del proveedor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Provider"
                }
              }
            }
          },
          "404": {
            "description": "Proveedor no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un proveedor existente",
        "tags": [
          "Providers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "telefono": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proveedor actualizado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un proveedor",
        "tags": [
          "Providers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proveedor eliminado"
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "summary": "Obtener todos los productos con paginación",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Buscar por nombre"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de productos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo producto",
        "tags": [
          "Products"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre",
                  "margenLocal",
                  "margenViaje",
                  "costoUnitarioReal"
                ],
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "margenLocal": {
                    "type": "number"
                  },
                  "margenViaje": {
                    "type": "number"
                  },
                  "costoUnitarioReal": {
                    "type": "number"
                  },
                  "stockDisponible": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Obtener un producto por ID",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del producto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un producto existente",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "margenLocal": {
                    "type": "number"
                  },
                  "margenViaje": {
                    "type": "number"
                  },
                  "costoUnitarioReal": {
                    "type": "number"
                  },
                  "stockDisponible": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Producto actualizado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un producto",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto eliminado"
          }
        }
      }
    },
    "/api/dashboard/daily": {
      "get": {
        "summary": "Obtiene métricas financieras del día y valor total del inventario",
        "tags": [
          "Dashboard"
        ],
        "responses": {
          "200": {
            "description": "Métricas financieras calculadas exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ingresos_brutos": {
                      "type": "number"
                    },
                    "utilidad_neta_dia": {
                      "type": "number"
                    },
                    "cantidad_operaciones": {
                      "type": "integer"
                    },
                    "fuga_por_redondeos": {
                      "type": "number"
                    },
                    "capital_invertido": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "vendedor",
              "cajero"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Provider": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "margenLocal": {
            "type": "number"
          },
          "margenViaje": {
            "type": "number"
          },
          "stockDisponible": {
            "type": "integer"
          },
          "costoUnitarioReal": {
            "type": "number"
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "tags": []
}